combat targeting
Refactored destination and enemy targeting
