--- stc_game/Assets/Scripts/Camera/TargetController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 76ed2970f21ef9a41bb4dee1481c8efe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
