+++ stc_game/Assets/Scripts/Targets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1ed47d98ee0cd084787170541f733c8d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
