+++ stc_game/Assets/Scripts/Targets/DestinationController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 15ae69d312145904098c713ff2c1226d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
